Revert "✏️ rename: repository 파일 위치 이동, table명 변경"
