+++ src/main/java/com/luckyGirls/ForYourNutrition/dao/jpa/repository/IRecommendJpaRepository.java
@@ -1,4 +1,4 @@
-package com.luckyGirls.ForYourNutrition.repository;
+package com.luckyGirls.ForYourNutrition.dao.jpa.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
