+++ src/main/java/com/luckyGirls/ForYourNutrition/service/IRecommendService.java
@@ -3,8 +3,8 @@ package com.luckyGirls.ForYourNutrition.service;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import com.luckyGirls.ForYourNutrition.repository.IRecommendJpaRepository;
-import com.luckyGirls.ForYourNutrition.repository.ItemJpaRepository;
+import com.luckyGirls.ForYourNutrition.dao.jpa.repository.IRecommendJpaRepository;
+import com.luckyGirls.ForYourNutrition.dao.jpa.repository.ItemJpaRepository;
 
 import lombok.RequiredArgsConstructor;
