+++ src/main/java/com/luckyGirls/ForYourNutrition/service/ItemService.java
@@ -3,7 +3,7 @@ package com.luckyGirls.ForYourNutrition.service;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import com.luckyGirls.ForYourNutrition.repository.ItemJpaRepository;
+import com.luckyGirls.ForYourNutrition.dao.jpa.repository.ItemJpaRepository;
 import com.luckyGirls.ForYourNutrition.domain.Item;
 import com.luckyGirls.ForYourNutrition.dto.response.ItemGetResponse;
